refactor(eventproc): publish EventProducer/EventProcessor to pkg/eventproc (SRD-012 M1b)
Second leaf of the ADR-012 layering relocation. Carve the two node-facing
event contracts out of internal/eventproc into the public pkg/eventproc;
EventHub / EventWaiter / EventWaiterState and the eventhub implementation stay
internal.

internal/eventproc re-exports EventProducer/EventProcessor as type aliases to
the public package, so its EventHub/EventWaiter and every internal consumer
(eventhub, waiters, instance, track, thresher) keep referring to them
unqualified and untouched. Only the model file events/event.go re-points to
pkg/eventproc — the layering goal (model now imports neither internal/eventproc
nor internal/interactor). .mockery.yaml split: pkg/eventproc mocks
EventProducer/EventProcessor, internal/eventproc mocks EventHub/EventWaiter;
both co-generate into mockeventproc so test call-sites are unchanged.

Behaviour-preserving. Implements SRD-012 v.1 FR-3 (and completes M1).
